New main activity, renamed some files, modified manifest
