+++ restaurateur/src/main/res/layout/activity_main.xml
@@ -1,75 +1,47 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="20dp"
-    android:layout_marginStart="16dp"
-    android:layout_marginEnd="16dp">
 
-    <de.hdodenhof.circleimageview.CircleImageView
-        xmlns:ristoratore="http://schemas.android.com/apk/res-auto"
-        android:id="@+id/avatar"
-        android:layout_marginTop="20dp"
-        android:layout_gravity="center"
-        android:layout_width="160dp"
-        android:layout_height="160dp"
-        android:contentDescription="@string/photo"
-        android:src="@drawable/ic_launcher_foreground"
-        ristoratore:civ_border_color="#d1b1b1"
-        ristoratore:civ_fill_color="@color/colorAccent"
-        android:foregroundGravity="center"
-        />
-
-    <TextView
-        android:id="@+id/name_text"
+    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="40dp"
-        android:hint="@string/cust_name"
-        android:textAppearance="@android:style/TextAppearance.Material.Large" />
+        android:layout_height="match_parent"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:orientation="vertical"
+        android:padding="20dp">
 
-    <TextView
-        android:id="@+id/address_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:hint="@string/cust_addr"
-        android:textAppearance="@android:style/TextAppearance.Material.Large" />
 
-    <TextView
-        android:id="@+id/tel_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:hint="@string/cust_tel"
-        android:textAppearance="@android:style/TextAppearance.Material.Large" />
+        <Button
+            android:id="@+id/profile_btn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginBottom="20dp"
+            android:text="Profile"
+            tools:layout_editor_absoluteX="152dp"
+            tools:layout_editor_absoluteY="268dp" />
 
-    <TextView
-        android:id="@+id/mail_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:hint="@string/cust_mail"
-        android:textAppearance="@android:style/TextAppearance.Material.Large" />
+        <Button
+            android:id="@+id/menu_btn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginTop="20dp"
+            android:text="Daily Menu"
+            tools:layout_editor_absoluteX="152dp"
+            tools:layout_editor_absoluteY="268dp" />
 
-    <TextView
-        android:id="@+id/hour_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:hint="@string/cust_hour"
-        android:textAppearance="@android:style/TextAppearance.Material.Large" />
-
-    <TextView
-        android:id="@+id/info_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:hint="@string/cust_info"
-        android:textAppearance="@android:style/TextAppearance.Material.Large" />
-</LinearLayout>
+        <Button
+            android:id="@+id/order_btn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginTop="40dp"
+            android:text="Orders"
+            tools:layout_editor_absoluteX="161dp"
+            tools:layout_editor_absoluteY="537dp" />
+    </LinearLayout>
 </ScrollView>
